Create GFDataLoadingVC and make some screens subclass of it, so only they can display data related views
